Apply MVC to advanced customizations settings
